+++ src/typescript-worker/util.ts
@@ -174,13 +174,63 @@ export const charToLineAndChar = (
   return {line, character};
 };
 
+/**
+ * The "exports" field of a package.json.
+ *
+ * See https://nodejs.org/api/packages.html#packages_exports.
+ */
+export type PackageExports =
+  | PackageExportsTarget
+  | PackageExportsPathOrConditionMap;
+
+/**
+ * The export result for some path or condition.
+ */
+export type PackageExportsTarget =
+  // A concrete path.
+  | PackageExportsTargetPath
+  // Condition maps can be nested.
+  | PackageExportsConditionMap
+  // The first valid target in an array wins.
+  | PackageExportsTarget[]
+  // An explicit "not found".
+  | null;
+
+/**
+ * A concrete resolved path (e.g. "./lib/foo.js").
+ */
+export type PackageExportsTargetPath = string;
+
+/**
+ * Map from a path or condition to a target.
+ */
+export type PackageExportsPathOrConditionMap = {
+  [pathOrCondition: string]: PackageExportsTarget;
+};
+
+/**
+ * Map from a condition to a target.
+ *
+ * Note this is technically the same type as PackageExportsPathOrConditionMap,
+ * but it's distinguished for clarity because "path" keys are only allowed in
+ * the top-level of the "exports" object.
+ */
+export type PackageExportsConditionMap = {
+  [condition: string]: PackageExportsTarget;
+};
+
 export interface PackageJson {
   version?: string;
   main?: string;
+  exports?: PackageExports;
   module?: string;
   types?: string;
   typings?: string;
   dependencies?: {[key: string]: string};
+}
+
+export interface PackageJsonWithExports extends PackageJson {
+  exports: PackageExports;
 }
 
 /**